Optimization
Removed smn_dict, it is impossible to use github.com/ProtossGenius/smn_dict when using it in other files, this file should be generated by the user.

+++ test/smn_net_rpc/test.go
@@ -46,7 +46,7 @@ func AccpterRun(adapter smn_rpc.MessageAdapterItf) {
 		msg, err := adapter.ReadCall()
 		check(err)
 		dict, res, err := rpcSvr.OnMessage(msg, adapter.GetConn())
-		adapter.WriteRet(dict, res, err)
+		adapter.WriteRet(int32(dict), res, err)
 	}
 }
